score test: do score test before GP fitting
the previous strategy may have been generating false-positives due to
the score test being performed with the fitted kernel. The current
workflow is to first run the score test on multiple kernels, and then
analyze the detected spatially variable genes in more detail with GPs.
Based on extensive benchmarking, the Cauchy combination seems to perform
best in terms of statistical power, so it is the default. If CPU time is
a bottleneck, the omnibus test can be used instead. It may lose some
power, but its runtime is independent of the number of kernels to be
tested. In the default configuration, it is 10x faster than the Cauchy
combination.

This commit also ports the score test to use Tensorflow, so it can run
on GPUs. This gives a 100x speed-up compared to CPU.