port AEH to tensorflow with Dirichlet Process prior
Originally, it was planned to also fit lengthscales for each kernel.
This is why a split optimization strategy is implemented, with one group
of parameters optimized with BFGS and another group of parameters
calculated from the former with mean field updates. However, it turns
out that with this strategy, the kernel lengthscale is always optimized
towards 0, such that all variation is covered by noise variance, which
is not what we want. Thus, this final version is again using fixed
lengthscales, as in SpatialDE1, but offers the user the option to
specify multiple lengthscales.