tissue segmentation: adjust alpha hyperprior
this matches the hyperprior in AEH and yiels better results. The reason
for this is that using 1e-6, 1e-6 as hyperprior paramters results in a
Gamma distribution with mean 1 and very large variance, meaning that
sampling from this distribution produces almost exclusively 0s with some
extremely large values. Simulations reveal that sampling from a
Beta(1,g) distribution, where g are the Gamma samples yields exclusively
1s. In other words, eta_1=1e-6 and eta_2=1e-6 is actually a very strong
prior for only 1 class